Move all extensions related code into extensions module
