+++ src/extensions.rs
@@ -1,15 +1,128 @@
 //! X.509 Extensions objects and types
 
-use crate::time::der_to_utctime;
-use crate::{ASN1Time, ReasonCode, X509Name};
+use std::collections::HashMap;
+use std::fmt;
+
 use der_parser::ber::*;
+use der_parser::der::{
+    der_read_element_header, parse_der_bool, parse_der_octetstring, parse_der_oid,
+};
+use der_parser::error::BerResult;
 use der_parser::oid::Oid;
 use nom::combinator::{all_consuming, complete, map_opt, map_res, opt};
 use nom::multi::{many0, many1};
+use nom::{exact, Err};
 use num_bigint::BigUint;
 use oid_registry::*;
-use std::collections::HashMap;
-use std::fmt;
+
+use crate::error::{X509Error, X509Result};
+use crate::time::der_to_utctime;
+use crate::{ASN1Time, ReasonCode, X509Name};
+
+#[derive(Debug, PartialEq)]
+pub struct X509Extension<'a> {
+    /// OID describing the extension content
+    pub oid: Oid<'a>,
+    /// Boolean value describing the 'critical' attribute of the extension
+    ///
+    /// An extension includes the boolean critical, with a default value of FALSE.
+    pub critical: bool,
+    /// Raw content of the extension
+    pub value: &'a [u8],
+    pub(crate) parsed_extension: ParsedExtension<'a>,
+}
+
+impl<'a> X509Extension<'a> {
+    /// Parse a DER-encoded X.509 extension
+    ///
+    /// X.509 extensions allow adding attributes to objects like certificates or revocation lists.
+    ///
+    /// Each extension in a certificate is designated as either critical or non-critical.  A
+    /// certificate using system MUST reject the certificate if it encounters a critical extension it
+    /// does not recognize; however, a non-critical extension MAY be ignored if it is not recognized.
+    ///
+    /// Each extension includes an OID and an ASN.1 structure.  When an extension appears in a
+    /// certificate, the OID appears as the field extnID and the corresponding ASN.1 encoded structure
+    /// is the value of the octet string extnValue.  A certificate MUST NOT include more than one
+    /// instance of a particular extension.
+    ///
+    /// This function parses the global structure (described above), and will return the object if it
+    /// succeeds. During this step, it also attempts to parse the content of the extension, if known.
+    /// The returned object has a
+    /// [parsed_extension](x509/struct.X509Extension.html#method.parsed_extension) method. The returned
+    /// enum is either a known extension, or the special value `ParsedExtension::UnsupportedExtension`.
+    ///
+    /// <pre>
+    /// Extension  ::=  SEQUENCE  {
+    ///     extnID      OBJECT IDENTIFIER,
+    ///     critical    BOOLEAN DEFAULT FALSE,
+    ///     extnValue   OCTET STRING  }
+    /// </pre>
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// # use x509_parser::{X509Extension, extensions::ParsedExtension};
+    /// #
+    /// static DER: &[u8] = &[
+    ///    0x30, 0x1D, 0x06, 0x03, 0x55, 0x1D, 0x0E, 0x04, 0x16, 0x04, 0x14, 0xA3, 0x05, 0x2F, 0x18,
+    ///    0x60, 0x50, 0xC2, 0x89, 0x0A, 0xDD, 0x2B, 0x21, 0x4F, 0xFF, 0x8E, 0x4E, 0xA8, 0x30, 0x31,
+    ///    0x36 ];
+    ///
+    /// # fn main() {
+    /// let res = X509Extension::from_der(DER);
+    /// match res {
+    ///     Ok((_rem, ext)) => {
+    ///         println!("Extension OID: {}", ext.oid);
+    ///         println!("  Critical: {}", ext.critical);
+    ///         let parsed_ext = ext.parsed_extension();
+    ///         assert!(*parsed_ext != ParsedExtension::UnsupportedExtension);
+    ///         if let ParsedExtension::SubjectKeyIdentifier(key_id) = parsed_ext {
+    ///             assert!(key_id.0.len() > 0);
+    ///         } else {
+    ///             panic!("Extension has wrong type");
+    ///         }
+    ///     },
+    ///     _ => panic!("x509 extension parsing failed: {:?}", res),
+    /// }
+    /// # }
+    /// ```
+    pub fn from_der(i: &'a [u8]) -> X509Result<Self> {
+        parse_ber_sequence_defined_g(|_, i| {
+            let (i, oid) = map_res(parse_der_oid, |x| x.as_oid_val())(i)?;
+            let (i, critical) = der_read_critical(i)?;
+            let (i, value) = map_res(parse_der_octetstring, |x| x.as_slice())(i)?;
+            let (i, parsed_extension) = crate::extensions::parser::parse_extension(i, value, &oid)?;
+            let ext = X509Extension {
+                oid,
+                critical,
+                value,
+                parsed_extension,
+            };
+            Ok((i, ext))
+        })(i)
+        .map_err(|_| X509Error::InvalidExtensions.into())
+    }
+
+    pub fn new(
+        oid: Oid<'a>,
+        critical: bool,
+        value: &'a [u8],
+        parsed_extension: ParsedExtension<'a>,
+    ) -> X509Extension<'a> {
+        X509Extension {
+            oid,
+            critical,
+            value,
+            parsed_extension,
+        }
+    }
+
+    /// Return the extension type or `UnsupportedExtension` if the extension is not implemented.
+    pub fn parsed_extension(&self) -> &ParsedExtension<'a> {
+        &self.parsed_extension
+    }
+}
 
 #[derive(Debug, PartialEq)]
 pub enum ParsedExtension<'a> {
@@ -798,6 +911,57 @@ pub(crate) mod parser {
         let (rest, num) = map_opt(parse_der_integer, |obj| obj.as_biguint())(i)?;
         Ok((rest, ParsedExtension::CRLNumber(num)))
     }
+}
+
+/// Extensions  ::=  SEQUENCE SIZE (1..MAX) OF Extension
+pub(crate) fn parse_extension_sequence(i: &[u8]) -> X509Result<Vec<X509Extension>> {
+    parse_ber_sequence_defined_g(|_, a| all_consuming(many0(complete(X509Extension::from_der)))(a))(
+        i,
+    )
+}
+
+pub(crate) fn extensions_sequence_to_map<'a>(
+    i: &'a [u8],
+    v: Vec<X509Extension<'a>>,
+) -> X509Result<'a, HashMap<Oid<'a>, X509Extension<'a>>> {
+    let mut extensions = HashMap::new();
+    for ext in v.into_iter() {
+        if extensions.insert(ext.oid.clone(), ext).is_some() {
+            // duplicate extensions are not allowed
+            return Err(Err::Failure(X509Error::DuplicateExtensions));
+        }
+    }
+    Ok((i, extensions))
+}
+
+pub(crate) fn parse_extensions(
+    i: &[u8],
+    explicit_tag: BerTag,
+) -> X509Result<HashMap<Oid, X509Extension>> {
+    if i.is_empty() {
+        return Ok((i, HashMap::new()));
+    }
+
+    match der_read_element_header(i) {
+        Ok((rem, hdr)) => {
+            if hdr.tag != explicit_tag {
+                return Err(Err::Error(X509Error::InvalidExtensions));
+            }
+            let (rem, list) = exact!(rem, parse_extension_sequence)?;
+            extensions_sequence_to_map(rem, list)
+        }
+        Err(_) => Err(X509Error::InvalidExtensions.into()),
+    }
+}
+
+fn der_read_critical(i: &[u8]) -> BerResult<bool> {
+    // parse_der_optional!(i, parse_der_bool)
+    let (rem, obj) = opt(parse_der_bool)(i)?;
+    let value = obj
+        .map(|o| o.as_bool().unwrap_or_default()) // unwrap cannot fail, we just read a bool
+        .unwrap_or(false) // default critical value
+        ;
+    Ok((rem, value))
 }
 
 #[cfg(test)]
